macOS: Don't call [NSOpenGLContext update] for every frame
Calling update has a cost, and should only be done when the drawable
object changes size or location. Instead of calling update each time
makeCurrent is called, we listen for the appropriate notifications,
limiting the number of update calls significantly.

The code has also been refactored to get rid of the m_activeWindow
member, as the active window can be tracked through the context's
drawable object property.

There is also no need to clear the drawable when a window is hidden,
so the hook into QCocoaWindow can be removed.

The QPlatformNativeInterface hook is internal and can safely be removed.

Task-number: QTBUG-63572
Change-Id: I70e3267f47882e151144bd36a50abe906164429a
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>