Glib dispatcher: use default priority for timers
For some reason, the GLib event dispatcher was lowering the priority for
timers compared to other events like for example socket events.

the IDLE priority is for "Do this later when you are not busy", and that
is wrong for generic timer events.

See also https://bugs.kde.org/show_bug.cgi?id=230184

Task-number: QTBUG-48344
Change-Id: Id337a8f0c7fad3ebd5e5fa49d7831a0001ea9010
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: David Faure <david.faure@kdab.com>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>