Add support for additional prefix paths
That is, paths that are searched for dependencies too, next to the Qt
install prefix.

So far this only covers library dependencies, dependency meta data or
QML imports are subject to subsequent patches.

src/tools/androiddeployqt/main.cpp was moved from qttools into
qtbase in 5.11. So re-apply 14013829 in qttools here.

Change-Id: I5649e518cfa3deeac429b4142e6292cb514ef104
Reviewed-by: Volker Krause <volker.krause@kdab.com>