Improve detection and handling of unsupported Apple platforms
The application name wasn't always printed, so we try try a few
more possibilities before falling back to the process name.

We also run the check as early as possible, instead of relying
on a QCoreApplication.

We do not have to provide a dialog to the user, as macOS will do
this for us if the application is launched from Finder.

Change-Id: Ifbec86946d60294806364e08964852fd4b74ff56
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>